Move DynamicSoundEffectInstanceManager
