+++ MonoGame.Framework/Audio/DynamicSoundEffectInstance.XAudio.cs
@@ -1,4 +1,8 @@
-﻿using System;
+﻿// MonoGame - Copyright (C) The MonoGame Team
+// This file is subject to the terms and conditions defined in
+// file 'LICENSE.txt', which is part of this source code package.
+
+using System;
 using System.Collections.Generic;
 using SharpDX;
 using SharpDX.Multimedia;
